+++ BetterCrafting/Models/CraftingStation.cs
@@ -1,9 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 namespace Leclair.Stardew.BetterCrafting.Models;
 
 public class CraftingStation : ICraftingStation {
@@ -20,8 +14,12 @@ public class CraftingStation : ICraftingStation {
 
 	public bool IsCooking { get; set; } = false;
 
-	public string[] Recipes { get; set; } = Array.Empty<string>();
+	public string[] Recipes { get; set; } = [];
 
 	public Category[]? Categories { get; set; } = null;
+
+
+
+
 
 }