+++ Common/Crafting/IPerformCraftEvent.cs
@@ -1,5 +1,7 @@
 #nullable enable
 
+using System.Collections.Generic;
+
 using StardewValley;
 using StardewValley.Menus;
 
@@ -56,5 +58,26 @@ public interface IGlobalPerformCraftEvent : IPerformCraftEvent {
 	/// The recipe being crafted.
 	/// </summary>
 	IRecipe Recipe { get; }
+
+}
+
+/// <summary>
+/// A further extended IGlobalPerformCraftEvent subclass that includes
+/// a list of the matching ingredients being consumed. This is necessary
+/// because adding this to the existing model would break Pintail's
+/// proxying, for some reason.
+/// </summary>
+public interface IGlobalPerformCraftEventV2 : IGlobalPerformCraftEvent {
+
+	/// <summary>
+	/// A dictionary for looking up matching ingredient items.
+	/// If an <see cref="IIngredient"/> does not have an entry in this
+	/// list, then that ingredient does not support the ingredient
+	/// matching feature.
+	///
+	/// You may remove an item from the list to prevent it from
+	/// being consumed.
+	/// </summary>
+	IReadOnlyDictionary<IIngredient, List<Item>> MatchingItems { get; }
 
 }